Create CODE-1 for max30100 pulse oximeter heart rate sensor module with arduino
This code will initialize the MAX30100 module and then continuously read data from it in the loop() function. The pulseOximeter.read() function returns a PulseOximeterReading object that contains the heart rate and other data. In this example, we are just printing the heart rate to the serial monitor.

You can modify this code to do other things with the data from the MAX30100 module, such as controlling an LED or displaying the data on an LCD screen.